requirements.txt
asgiref==3.8.1
attrs==23.2.0
az-iranian-bank-gateways==2.0.5
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.6
django-admin-persian-fonts==0.2
django-ckeditor==6.6.0
django-cleanup==8.1.0
django-js-asset==2.2.0
django-modeltranslation==0.19.5
django-mptt==0.16.0
django-parler==2.3
django-phonenumber-field==8.0.0
django-user-language-middleware==0.0.3
idna==3.7
isodate==0.6.1
lxml==5.2.2
phonenumbers==8.13.39
phonenumberslite==8.13.39
pillow==10.3.0
platformdirs==4.2.2
pycryptodome==3.20.0
python-gettext==5.0
pytz==2024.1
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.2
zeep==4.2.1

persiantools~=4.2.0